+++ flutter_readium/example/lib/pages/bookshelf.page.dart
@@ -194,6 +194,9 @@ class BookshelfPageState extends State<BookshelfPage> {
 
   String _bookFormatFromConformsTo(Publication pub) {
     if (pub.conformsToReadiumEbook) {
+      if (pub.isAudioBook) {
+        return 'Ebook with audio';
+      }
       return 'Ebook';
     } else if (pub.conformsToReadiumAudiobook) {
       return 'Audiobook';
@@ -253,7 +256,7 @@ class BookshelfPageState extends State<BookshelfPage> {
                 ],
               ),
               // remove the if when books loaded from asset can be deleted
-              if (!_identifiersFromAsset.contains(publication.identifier))
+              if (!kIsWeb && !_identifiersFromAsset.contains(publication.identifier))
                 IconButton(
                   icon: Icon(Icons.delete, color: Colors.red),
                   onPressed: () async {